Clear chain configs & saved pages in Clear Storage

+++ src/pages/settings/components/FormSettings.tsx
@@ -201,7 +201,9 @@ export const FormSettings = () => {
     setIsButtonLoading(true);
     const deleteDBRequest = indexedDB.deleteDatabase('NeDB');
     deleteDBRequest.onsuccess = () => {
-      localStorage.clear();
+      if(localStorage) {
+        localStorage.clear();
+      }
       setTimeout(() => {
         ipcRenderer.send('restart_app');
       }, 2000);
